+++ yesod-test/ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Fix quotes not matching in htmlContain* functions [#1768](https://github.com/yesodweb/yesod/pull/1768).
 * Add logging of the matches found of these functions [#1768](https://github.com/yesodweb/yesod/pull/1768).
+* Improved failure messages from `assertEq`. [#1767](https://github.com/yesodweb/yesod/pull/1767)
 
 ## 1.6.13